feat(docs): Switch to JSDOC docs not in the readme
- Improve a few docs here and there
- Rename the repo
- Remove the authors field in the package.json
