--- .docs/edit-warning.hbs
@@ -1,10 +1,0 @@
-<!-- STOP! WAIT! BEFORE YOU EDIT!
-
-README.md is a generated file. Please don't edit it!
-
-This file is generated from .docs/README.hbs and the JavaScript files themselves.
-See our contributing guide for more information.
-
-https://github.com/EmergingTechnologyAdvisors/node-serialport/blob/master/CONTRIBUTING.md#writing-docs
-
--->
