--- .docs/sig-name.hbs
@@ -1,15 +1,0 @@
-{{#if virtual}}*{{/if}}{{#with (parentObject)}}{{#if virtual}}*{{/if~}}{{/with~}}
-{{#if name}}{{#sig~}}
-{{{@depOpen}~}}
-{{{@codeOpen}~}}
-{{#if @prefix}}{{@prefix}} {{/if~}}
-{{@parent~}}
-{{@accessSymbol}}{{#if (isEvent)}}Event: "{{{name}}}"{{else}}{{{name}}}{{/if~}}
-{{#if @methodSign}}{{#if (isEvent)}} {{@methodSign}}{{else}}{{@methodSign}}{{/if}}{{/if~}}
-{{{@codeClose}~}}
-{{#if @returnSymbol}} {{@returnSymbol}}{{/if~}}
-{{#if @returnTypes}} {{>linked-type-list types=@returnTypes delimiter=" &#124; " }}{{/if~}}
-{{#if @suffix}} {{@suffix}}{{/if~}}
-{{{@depClose}~}}
-{{~/sig}}{{/if~}}
-{{#if virtual}}*{{/if}}{{#with (parentObject)}}{{#if virtual}}*{{/if~}}{{/with~}}
