+++ .gitignore
@@ -7,3 +7,4 @@ npm-debug.log
 package-lock.json
 Release/
 prebuilds/
+docs
