+++ .jsdoc.json
@@ -1,0 +1,34 @@
+{
+  "tags": {
+    "allowUnknownTags": true,
+    "dictionaries": ["jsdoc"]
+  },
+  "source": {
+    "include": ["lib", "README.md"],
+    "includePattern": ".+\\.(js|jsx)?$"
+  },
+  "plugins": [
+    "plugins/markdown"
+  ],
+  "markdown": {
+		"parser": "gfm",
+		"hardwrap": true,
+		"excludeTags": []
+	},
+  "templates": {
+    "cleverLinks": true,
+    "monospaceLinks": true
+  },
+  "opts": {
+    "template": "node_modules/docdash",
+    "encoding": "utf8",
+    "private": true,
+    "destination": "./docs/",
+    "recurse": true,
+    "_tutorials": "./examples"
+  },
+  "docdash": {
+    "static": true,
+    "sort": true
+  }
+}
