+++ .npmignore
@@ -27,3 +27,4 @@ node_modules
 npm-debug.log
 package-lock.json
 Release/
+docs
