+++ PUBLISHING.md
@@ -8,11 +8,10 @@ This can be checked in the .travis.yml file and appveyor.yml file. Within these 
 
 1. Bump up npm version in `package.json`
 2. Run `npm run changelog` and modify `CHANGELOG.md` if needed
-3. Update the `.docs/README.hbs` to reference this current version and to previous major version docs then regenerate docs `npm run docs`.
-4. Commit everything then generate new tags based on package.json version number with `git tag v6.0.0 -a` and include the change log in the tag's annotation. (beta `git tag v6.0.0-beta3 -a`)
-5. Push tags to Github with `git push --tags`
-6. Wait for the CI to publish all the binaries. Remove the content of the Github release message so the tag's text shows.
-7. `rm -rf package-lock.json node_modules build && npm install`
-8. When the entire matrix succeeds and all binaries exist run `npm publish` or `npm publish --tag beta`.
-9. Kick off the build matrix for either the master or beta branch on [serialport-test-pilot](https://travis-ci.org/j5js/serialport-test-pilot). It will install serialport from npm on a wide range of systems.
-10. Let everyone know 🎉
+3. Commit everything then generate new tags based on package.json version number with `git tag v6.0.0 -a` and include the change log in the tag's annotation. (beta `git tag v6.0.0-beta3 -a`)
+4. Push tags to Github with `git push --tags`
+5. Wait for the CI to publish all the binaries. Remove the content of the Github release message so the tag's text shows.
+6. `rm -rf package-lock.json node_modules build && npm install`
+7. When the entire matrix succeeds and all binaries exist run `npm publish` or `npm publish --tag beta`.
+8. Kick off the build matrix for either the master or beta branch on [serialport-test-pilot](https://travis-ci.org/j5js/serialport-test-pilot). It will install serialport from npm on a wide range of systems.
+9. Let everyone know 🎉
