+++ lib/bindings/base.js
@@ -1,8 +1,5 @@
 'use strict';
 const debug = require('debug')('serialport:bindings');
-/**
- * @module serialport
- */
 
 /**
  * @name module:serialport.Binding
@@ -18,7 +15,7 @@ const debug = require('debug')('serialport:bindings');
 
 /**
  * You never have to use `Binding` objects directly. SerialPort uses them to access the underlying hardware. This documentation is geared towards people who are making bindings for different platforms. This class can be inherited from to get type checking for each method.
- * @class
+ * @class BaseBinding
  * @param {object} options
  * @property {boolean} isOpen Required property. `true` if the port is open, `false` otherwise. Should be read-only.
  * @throws {TypeError} When given invalid arguments, a `TypeError` is thrown.
