+++ lib/bindings/poller.js
@@ -4,6 +4,10 @@ const logger = debug('serialport:poller');
 const EventEmitter = require('events');
 const FDPoller = require('bindings')('serialport.node').Poller;
 
+/**
+ * Enum of event values
+ * @enum {int}
+ */
 const EVENTS = {
   UV_READABLE: 1,
   UV_WRITABLE: 2,
@@ -35,13 +39,17 @@ function handleEvent(error, eventFlag) {
 /**
  * Polls unix systems for readable or writable states of a file or serialport
  */
-module.exports = class Poller extends EventEmitter {
+class Poller extends EventEmitter {
   constructor(fd) {
     logger('Creating poller');
     super();
     this.poller = new FDPoller(fd, handleEvent.bind(this));
   }
-
+  /**
+   * Wait for the next event to occur
+   * @param {string} Event ('readable'|'writable'|'disconnect')
+   * @param {function} callback
+   */
   once(event) {
     switch (event) {
       case 'readable':
@@ -57,6 +65,10 @@ module.exports = class Poller extends EventEmitter {
     return EventEmitter.prototype.once.apply(this, arguments);
   }
 
+  /**
+   * Ask the bindings to listen for an event
+   * @param {EVENTS} eventFlag
+   */
   poll(eventFlag) {
     eventFlag = eventFlag || 0;
 
@@ -73,13 +85,12 @@ module.exports = class Poller extends EventEmitter {
     this.poller.poll(eventFlag);
   }
 
+  /**
+   * Stop listening for events and cancel all outstanding listening with an error
+   */
   stop() {
     logger('Stopping poller');
     this.poller.stop();
-    this.emitCanceled();
-  }
-
-  emitCanceled() {
     const err = new Error('Canceled');
     err.canceled = true;
     this.emit('readable', err);
@@ -87,3 +98,7 @@ module.exports = class Poller extends EventEmitter {
     this.emit('disconnect', err);
   }
 };
+
+Poller.EVENTS = EVENTS;
+
+module.exports = Poller;
