+++ lib/index.js
@@ -1,15 +1,16 @@
 'use strict';
-
-/**
- * @module serialport
- * @copyright Chris Williams <chris@iterativedesigns.com>
- */
-
 const SerialPort = require('./serialport');
 const Binding = require('./bindings/auto-detect');
 const parsers = require('./parsers');
 
+/**
+ * @type {BaseBinding}
+ */
 SerialPort.Binding = Binding;
+
+/**
+ * @type {Parsers}
+ */
 SerialPort.parsers = parsers;
 
 module.exports = SerialPort;
