+++ lib/parsers/byte-length.js
@@ -2,7 +2,20 @@
 const Buffer = require('safe-buffer').Buffer;
 const Transform = require('stream').Transform;
 
-module.exports = class ByteLengthParser extends Transform {
+/**
+ * A transform stream that emits data as a buffer after a specific number of bytes are received.
+ * @extends Transform
+ * @example
+To use the `ByteLength` parser, provide the length of the number of bytes:
+```js
+const SerialPort = require('serialport');
+const ByteLength = SerialPort.parsers.ByteLength
+const port = new SerialPort('/dev/tty-usbserial1');
+const parser = port.pipe(new ByteLength({length: 8}));
+parser.on('data', console.log);
+```
+ */
+class ByteLengthParser extends Transform {
   constructor(options) {
     super(options);
     options = options || {};
@@ -36,3 +49,5 @@ module.exports = class ByteLengthParser extends Transform {
     cb();
   }
 };
+
+module.exports = ByteLengthParser;
