+++ lib/parsers/cctalk.js
@@ -1,8 +1,20 @@
 'use strict';
 const Transform = require('stream').Transform;
 const Buffer = require('safe-buffer').Buffer;
-
-module.exports = class CCTalkParser extends Transform {
+/**
+ * Parses the CCTalk protocol
+ * @extends Transform
+ * @example
+CCTalk Messages are emitted as buffers.
+```js
+const SerialPort = require('serialport');
+const CCTalk = SerialPort.parsers.CCTalk;
+const port = new SerialPort('/dev/ttyUSB0');
+const parser = port.pipe(new CCtalk());
+parser.on('data', console.log);
+```
+ */
+class CCTalkParser extends Transform {
   constructor() {
     super();
     this.array = [];
@@ -29,3 +41,5 @@ module.exports = class CCTalkParser extends Transform {
     cb();
   }
 };
+
+module.exports = CCTalkParser;
