+++ lib/parsers/delimiter.js
@@ -1,8 +1,20 @@
 'use strict';
 const Buffer = require('safe-buffer').Buffer;
 const Transform = require('stream').Transform;
-
-module.exports = class DelimiterParser extends Transform {
+/**
+ * A transform stream that emits data each time a byte sequence is received.
+ * @extends Transform
+ * @example
+To use the `Delimiter` parser, provide a delimiter as a string, buffer, or array of bytes:
+```js
+const SerialPort = require('serialport');
+const Delimiter = SerialPort.parsers.Delimiter;
+const port = new SerialPort('/dev/tty-usbserial1');
+const parser = port.pipe(new Delimiter({ delimiter: Buffer.from('EOL') }));
+parser.on('data', console.log);
+```
+ */
+class DelimiterParser extends Transform {
   constructor(options) {
     options = options || {};
     super(options);
@@ -36,3 +48,5 @@ module.exports = class DelimiterParser extends Transform {
     cb();
   }
 };
+
+module.exports = DelimiterParser;
