+++ lib/parsers/ready.js
@@ -1,8 +1,21 @@
 'use strict';
 const Buffer = require('safe-buffer').Buffer;
 const Transform = require('stream').Transform;
-
-module.exports = class ReadyParser extends Transform {
+/**
+ * A transform stream that waits for a sequence of "ready" bytes before emitting a ready event and emitting data events
+ * @extends Transform
+ * @example
+To use the `Ready` parser provide a byte start sequence. After the bytes have been received a ready event is fired and data events are passed through.
+```js
+const SerialPort = require('serialport');
+const Ready = SerialPort.parsers.Ready;
+const port = new SerialPort('/dev/tty-usbserial1');
+const parser = port.pipe(new Ready({ data: 'READY' }));
+parser.on('ready', () => console.log('the ready byte sequence has been received'))
+parser.on('data', console.log); // all data after READY is received
+```
+ */
+class ReadyParser extends Transform {
   constructor(options) {
     options = options || {};
     if (options.delimiter === undefined) {
@@ -45,3 +58,5 @@ module.exports = class ReadyParser extends Transform {
     cb();
   }
 };
+
+module.exports = ReadyParser;
