+++ lib/parsers/regex.js
@@ -1,7 +1,20 @@
 'use strict';
 const Transform = require('stream').Transform;
+/**
+ * A transform stream that uses a regular expression to split the incoming text upon.
+ * @extends Transform
+ * @example
+To use the `Regex` parser provide a regular expression to split the incoming text upon. Data is emitted as string controllable by the `encoding` option (defaults to `utf8`).
+```js
+const SerialPort = require('serialport');
+const Regex = SerialPort.parsers.Regex;
+const port = new SerialPort('/dev/tty-usbserial1');
+const parser = port.pipe(new Regex({ regex: /[\r\n]+/ }));
+parser.on('data', console.log);
+```
 
-module.exports = class RegexParser extends Transform {
+ */
+class RegexParser extends Transform {
   constructor(options) {
     const opts = Object.assign({
       encoding: 'utf8'
@@ -37,3 +50,5 @@ module.exports = class RegexParser extends Transform {
     cb();
   }
 };
+
+module.exports = RegexParser;
