+++ lib/serialport.js
@@ -1,10 +1,4 @@
 'use strict';
-
-/**
- * @module serialport
- * @copyright Chris Williams <chris@iterativedesigns.com>
- */
-
 const Buffer = require('safe-buffer').Buffer;
 const stream = require('stream');
 const util = require('util');
@@ -82,7 +76,7 @@ function allocNewReadPool(poolSize) {
 
 /**
  * Create a new serial port object for the `path`. In the case of invalid arguments or invalid options, when constructing a new SerialPort it will throw an error. The port will open automatically by default, which is the equivalent of calling `port.open(openCallback)` in the next tick. You can disable this by setting the option `autoOpen` to `false`.
- * @class
+ * @class SerialPort
  * @param {string} path - The system path of the serial port you want to open. For example, `/dev/tty.XXX` on Mac/Linux, or `COM1` on Windows.
  * @param {module:serialport~openOptions=} options - Port configuration options
  * @param {module:serialport~errorCallback=} openCallback - Called after a connection is opened. If this is not provided and an error occurs, it will be emitted on the port's `error` event. The callback will NOT be called if `autoOpen` is set to `false` in the `openOptions` as the open will not be performed.
