+++ package.json
@@ -2,15 +2,10 @@
   "name": "serialport",
   "version": "6.0.4",
   "description": "Node.js package to access serial ports. Linux, OSX and Windows. Welcome your robotic JavaScript overlords. Better yet, program them!",
-  "author": {
-    "name": "Chris Williams",
-    "email": "voodootikigod@gmail.com",
-    "url": "http://www.voodootikigod.com"
-  },
   "main": "lib",
   "repository": {
     "type": "git",
-    "url": "git://github.com/EmergingTechnologyAdvisors/node-serialport.git"
+    "url": "git://github.com/node-serialport/node-serialport.git"
   },
   "keywords": [
     "ccTalk",
@@ -34,12 +29,18 @@
   ],
   "maintainers": [
     {
+      "name": "Francis Gulotta",
+      "email": "wizard@roborooter.com",
+      "url": "https://www.roborooter.com"
+    },
+    {
       "name": "Jacob Rosenthal",
       "email": "jakerosenthal@gmail.com"
     },
     {
       "name": "Chris Williams",
-      "email": "voodootikigod@gmail.com"
+      "email": "voodootikigod@gmail.com",
+      "url": "http://www.voodootikigod.com"
     },
     {
       "name": "Joe Ferner",
@@ -52,10 +53,6 @@
     {
       "name": "Rob Giseburt",
       "email": "giseburt@gmail.com"
-    },
-    {
-      "name": "Francis Gulotta",
-      "email": "wizard@roborooter.com"
     }
   ],
   "dependencies": {
@@ -72,6 +69,7 @@
     "chai": "^4.0.2",
     "chai-subset": "^1.5.0",
     "conventional-changelog-cli": "^1.3.2",
+    "docdash": "^0.4.0",
     "eslint": "^4.7.2",
     "eslint-config-standard": "^10.2.1",
     "eslint-plugin-import": "^2.7.0",
@@ -79,7 +77,7 @@
     "eslint-plugin-promise": "^3.5.0",
     "eslint-plugin-standard": "^3.0.1",
     "istanbul": "^0.4.4",
-    "jsdoc-to-markdown": "^3.0.0",
+    "jsdoc": "^3.5.5",
     "mocha": "^4.0.0",
     "prebuild": "^6.2.1",
     "proxyquire": "^1.7.10",
@@ -97,8 +95,7 @@
   "scripts": {
     "arduino-test": "TEST_PORT=$(./bin/find-arduino.js) npm test",
     "changelog": "conventional-changelog -i CHANGELOG.md -p angular -s",
-    "docs": "jsdoc2md --no-cache -t .docs/README.hbs --partial .docs/sig-name.hbs --partial .docs/sig-link.hbs --partial .docs/edit-warning.hbs -r table --separators --name-format -f lib/* lib/bindings/* lib/parsers/* > README.md",
-    "docs:diff": "jsdoc2md --no-cache -t .docs/README.hbs --partial .docs/sig-name.hbs --partial .docs/sig-link.hbs --partial .docs/edit-warning.hbs -r table --separators --name-format -f lib/* lib/bindings/* lib/parsers/* | diff README.md - || (echo 'Docs out of date, run `npm run docs` and commit the new README.md' && false)",
+    "docs": "jsdoc -c ./.jsdoc.json ",
     "lint": "eslint lib test bin examples",
     "rebuild-all": "npm rebuild && node-gyp rebuild",
     "repl": "node bin/repl.js",
